+++ CHANGELOG.md
@@ -13,6 +13,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/). 
   - XP gain from magic ops reduced to 1 XP per strength spent
   - XP gain from exploration increased to 0.6 per acre (from 0.25)
   - Cooldown to change hero classes reverted to 96 hours (from 72)
+- New Hero Upgrades
+  - Lead from the Front: Experience gains from invasion and exploration are doubled, but experience cannot be gained from magic and espionage
+  - Lead from the Shadows: 	Experience gains from magic and espionage are doubled
 
 ## [1.47.1] - 2025-12-11
 ### Changed